Added utils script to handle functions and changed Square from function to class
